+++ src/main/java/com/example/QuanLiThanhVienWeb/Repositories/ThanhVienRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ThanhVienRepository extends CrudRepository<ThanhVien,Long> {
+public interface ThanhVienRepository extends CrudRepository<ThanhVien, Integer> {
 }
